\name{apples}
\alias{apples}
\docType{data}
\title{
apples from southern hemisphere data set
}
\description{
Sensory characterization and consumers preference for 12 variables of apples
}
\usage{
data(apples)
}
\format{
  A data frame with 12 observations described by hedonic scores given by a panel of 60 consumers and  by 43 sensory attributes.

}


\references{
Daillant-Spinnler, B, MacFie, H.J.H, Beyts, P.K., Hedderley, D. (1996). Relationships between perceived sensory properties and major preference directions of 12 variaties of apples from the southern hemisphere. Food Quality and Preference, 7(2), 113-126.

}
\examples{
  data(apples)
}
\keyword{datasets}
