% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARI.R
\name{ARI}
\alias{ARI}
\title{Adjusted Rand Index}
\usage{
ARI(x, y)
}
\arguments{
\item{x}{vector defining a partition.}

\item{y}{vector defining a partition of whose length is equal to the length of x.}
}
\value{
numeric
}
\description{
This function computes the Adjusted Rand Index
}
\examples{
x <- sample(1:2, 20, replace=TRUE)
y <- x
y[1:5] <- sample(1:2, 5, replace=TRUE)
ARI(x, y)
}
\references{
L. Hubert and P. Arabie (1985) Comparing Partitions, Journal of the Classification, 2, pp. 193-218.
}
