% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractors.R
\docType{methods}
\name{BIC}
\alias{BIC}
\alias{BIC,VSLCMresults-method}
\title{BIC criterion.}
\usage{
\S4method{BIC}{VSLCMresults}(object)
}
\arguments{
\item{object}{instance of \code{\linkS4class{VSLCMresults}}.}
}
\description{
This function gives the BIC criterion (according to the formula log-likelihood - log(n)*npar/2, where npar represents the number of parameters in the fitted model and n represents the sample size)  for an instance of \code{\linkS4class{VSLCMresults}}.
}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. Annals of Statistics, 6(2), 461–464.
}
