% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractors.R
\docType{methods}
\name{coefficients}
\alias{coefficients}
\alias{coefficients,VSLCMresults-method}
\title{Extract the parameters}
\usage{
\S4method{coefficients}{VSLCMresults}(object)
}
\arguments{
\item{object}{instance of \code{\linkS4class{VSLCMresults}}.}
}
\description{
This function returns an instance of class \code{\linkS4class{VSLCMparam}} which contains the model parameters.
}
