% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractors.R
\docType{methods}
\name{fitted}
\alias{fitted}
\alias{fitted,VSLCMresults-method}
\title{Extract the partition or the probabilities of classification}
\usage{
\S4method{fitted}{VSLCMresults}(object, type = "partition")
}
\arguments{
\item{object}{instance of \code{\linkS4class{VSLCMresults}}.}

\item{type}{the type of prediction: probability of classification (probability) or the partition (partition)}
}
\description{
This function returns the partition among the observations of an instance of \code{\linkS4class{VSLCMresults}}.
}
