\docType{data}
\name{heart}
\alias{heart}
\title{Mixed data : Absence (1) or presence (2) of heart disease}
\format{A data frame with 270 observations on the following 12 variables.

\describe{

  \item{\code{Age}}{age}
  \item{\code{Sex}}{sex}
  \item{\code{ChestPainType}}{chest pain type  (4 values)  }
  \item{\code{RestBloodPressure}}{resting blood pressure}
  \item{\code{SerumCholestoral}}{serum cholestoral in mg/dl}
  \item{\code{FastingBloodSugar}}{fasting blood sugar > 120 mg/dl}
  \item{\code{ResElectrocardiographic}}{resting electrocardiographic results  (values 0,1,2) }
  \item{\code{MaxHeartRate}}{maximum heart rate achieved  }
  \item{\code{ExerciseInduced}}{exercise induced angina}
  \item{\code{Slope}}{the slope of the peak exercise ST segment }
  \item{\code{MajorVessels}}{number of major vessels (0-3) colored by flourosopy  }
  \item{\code{Thal}}{thal: 3 = normal; 6 = fixed defect; 7 = reversable defect }
  \item{\code{Class}}{Variable to be predicted: Absence (1) or presence (2) of heart disease}
}}
\source{
Gavin Brown. Diversity in Neural Network Ensembles. The University of Birmingham. 2004.
}
\description{
This database contains 13 attributes.
}
\examples{
data(heart)
}
\keyword{datasets}

