% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractors.R
\name{ICL}
\alias{ICL}
\title{ICL criterion}
\usage{
ICL(object)
}
\arguments{
\item{object}{\code{\linkS4class{VSLCMresults}}}
}
\description{
This function gives the ICL criterion for an instance of \code{\linkS4class{VSLCMresults}}.
}
\examples{
# Data loading:
data(heart)

# Cluster analysis without variable selection
res <- VarSelCluster(heart[,-13], 2, vbleSelec = FALSE)

# Get the ICL value
ICL(res)

}
\references{
Biernacki, C., Celeux, G., and Govaert, G. (2000). Assessing a mixture model for clustering with the integrated completed likelihood. IEEE transactions on pattern analysis and machine intelligence, 22(7), 719-725.
}
