% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractors.R
\name{MICL}
\alias{MICL}
\title{MICL}
\usage{
MICL(object)
}
\arguments{
\item{object}{\code{\linkS4class{VSLCMresults}}}
}
\description{
This function gives the MICL criterion for an instance of \code{\linkS4class{VSLCMresults}}.
}
\examples{
\dontrun{
data("heart")
z <- heart[,"Class"]
x <- heart[,-13]
object <- VarSelCluster(x, 2, vbleSelec = TRUE, crit.varsel = "MICL")
MICL(object)
}
}
\references{
Marbac, M. and Sedki, M. (2017). Variable selection for model-based clustering using the integrated completed-data likelihood. Statistics and Computing, 27 (4), 1049-1063.
}
