% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,VSLCMresults,character,ANY-method}
\alias{plot}
\alias{plot,VSLCMresults,ANY-method}
\title{This function draws information about an instance of \code{\linkS4class{VSLCMresults}}.}
\usage{
\S4method{plot}{VSLCMresults,character}(x, y, type)

\S4method{plot}{VSLCMresults,ANY}(x, type, ylim = c(1, x@data@d))
}
\arguments{
\item{x}{instance of  \code{\linkS4class{VSLCMresults}}.}

\item{y}{character. Must be the name of a variable in the analyzed data}

\item{...}{Additional argument list that might not ever be used.}

\item{x}{instance of  \code{\linkS4class{VSLCMresults}}.}

\item{...}{Additional argument list that might not ever be used.}
}
\description{
This function draws information about an instance of \code{\linkS4class{VSLCMresults}}.

This function draws information about an instance of \code{\linkS4class{VSLCMresults}}.
}
