\docType{data}
\name{SimulHetero}
\alias{SimulHetero}
\title{SimulHetero data}
\format{A simulated data set.
\describe{
  \item{\code{x}}{he whole data set}
  \item{\code{z}}{the component membership}
  \item{\code{xmiss}}{data set with missing values}
}
}
\description{
This data set is drawn from a bi-component mixture model.
}
\examples{
data(SimulHetero)
}
\keyword{datasets}

