\encoding{utf8}
\docType{data}
\name{stones}
\alias{stones}
\title{Stones data set}
\format{A data frame with 79 rows and 8 columns.}
\description{
The data set concerns the earliest history
of mankind. Prehistoric men created the desired shape of a stone tool
by striking on a raw stone, thus splitting off flakes, the waste
products of the crafting process. Archaelogists do not find many tools,
but they do find flakes. The data set is about these flakes. Its rows do
not stand for single flakes but for whole inventories of them. 
}
\arguments{
\item{LBI}{Length-breadth index of the striking platform}
\item{RTI}{Relative-thickness index of the striking platform}
\item{WDI}{Width-depth index of the striking platform}
\item{FLA}{Flaking angle (the angle between the striking platform and the splitting surface)}
\item{PSF}{Platform primery}
\item{FSF}{Platform facetted}
\item{ZDF1}{Dorsal surface totally worked}
\item{PROZD}{Proportion of worked dorsal surface (continuous)}
}
\examples{
data("stones")
}
\references{
Thomas Weber (2009): The Lower/middle palaeolithic transition - is
there a Lower/middle palaeolithic transition?
Preistoria Alpina 44: 1-6. 
}
\keyword{datasets}
