% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarSelConvert.R
\name{VarSelConvert}
\alias{VarSelConvert}
\title{Convert outputs of old versions of VarSelLCM}
\usage{
VarSelConvert(obj)
}
\arguments{
\item{obj}{an instance returned by function \link{VarSelCluster}.}
}
\description{
This function is used to convert output obtained by old versions (before 2.1.0) if VarSelLCM
}
\examples{
\dontrun{
##' # Data loading
data("heart")
# Clustering en 2 classes
heart[1,1] <- NA
results <- VarSelCluster(heart[,-13], 2)
# Here, it is necessary only if results was provided by an older version of the VarSelLCM
update <- VarSelConvert(results)
}

}
