% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarSelShiny.R
\name{VarSelShiny}
\alias{VarSelShiny}
\title{Shiny app for analyzing results from VarSelCluster}
\usage{
VarSelShiny(X)
}
\arguments{
\item{X}{an instance of \linkS4class{VSLCMresults} returned by function \link{VarSelCluster}.}
}
\description{
Shiny app for analyzing results from VarSelCluster
}
\examples{
\dontrun{
# Data loading
data("heart")
# Clustering en 2 classes
results <- VarSelCluster(heart[,-13], 2)
# Opening Shiny application to easily see the results
VarSelShiny(results)
}

}
