% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{VSLCMmodel-class}
\alias{VSLCMmodel-class}
\title{Constructor of \code{\linkS4class{VSLCMmodel}} class}
\description{
\describe{
  \item{g}{numeric. Number of components.}
  \item{omega}{logical. Vector indicating if each variable is irrelevant (1) or not (0) to the clustering.}
  \item{names.relevant}{character. Names of the relevant variables.}
  \item{names.irrelevant}{character. Names of the irrelevant variables.}
}
}
\examples{
  getSlots("VSLCMmodel")

}
