% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\docType{class}
\name{VSLCMparamCategorical-class}
\alias{VSLCMparamCategorical-class}
\title{Constructor of \code{\linkS4class{VSLCMparamCategorical}} class}
\description{
\describe{
  \item{pi}{numeric. Proportions of the mixture components.}
  \item{alpha}{list. Parameters of the multinomial distributions.}
}
}
\examples{
  getSlots("VSLCMparamCategorical")

}
