\name{VSLCMparameters-class}
\docType{class}
\alias{VSLCMparameters-class}

\title{Class \code{"VSLCMparameters"}}
\description{
This class contains the parameters
}



\section{Slots}{
  \describe{
    \item{\code{proportions}:}{Vector of the class proportions}
    \item{\code{means}:}{Matrix of the means where row correspond to classes and column to variables}
    \item{\code{variances}:}{Matrix of the variances where row correspond to classes and column to variables}
  }
}


\examples{
showClass("VSLCMparameters")
}

\keyword{classes}
