\name{VSLCMparamCategorical-class}
\docType{class}
\alias{VSLCMparamCategorical-class}

\title{Class \code{"VSLCMparamCategorical"}}
\description{
This class contains the paramCategorical
}



\section{Slots}{
  \describe{
    \item{\code{pi}:}{proportions}
    \item{\code{alpha}:}{List containing the probabilities ordered by variables.}
   
  }
}


\examples{
showClass("VSLCMparamCategorical")
}

\keyword{classes}
