\name{VSLCMparamContinuous-class}
\docType{class}
\alias{VSLCMparamContinuous-class}

\title{Class \code{"VSLCMparamContinuous"}}
\description{
This class contains the model information
}
\section{Slots}{
  \describe{
    \item{\code{pi}:}{proportions.}
    \item{\code{mu}:}{means.}
    \item{\code{sd}:}{standart deviations.}
  }
}
\examples{
showClass("VSLCMparamContinuous")
}
\keyword{classes}
