\name{VSLCMparamInteger-class}
\docType{class}
\alias{VSLCMparamInteger-class}

\title{Class \code{"VSLCMparamInteger"}}
\description{
This class contains the model information
}
\section{Slots}{
  \describe{
    \item{\code{pi}:}{proportions.}
    \item{\code{lambda}:}{means.}
  }
}
\examples{
showClass("VSLCMparamInteger")
}
\keyword{classes}
