% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\docType{class}
\name{VSLCMparamInteger-class}
\alias{VSLCMparamInteger-class}
\title{Constructor of \code{\linkS4class{VSLCMparamInteger}} class}
\description{
\describe{
  \item{pi}{numeric. Proportions of the mixture components.}
  \item{lambda}{matrix. Mean for each component (column) and each variable (row).}
}
}
\examples{
  getSlots("VSLCMparamInteger")

}
