% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSLCMGrlClasses.R
\docType{class}
\name{VSLCMparamMixed-class}
\alias{VSLCMparamMixed-class}
\title{Constructor of [\code{\linkS4class{VSLCMparamMixed}}] class}
\description{
\describe{
  \item{pi}{numeric. Proportions of the mixture components.}
  \item{paramContinuous}{\linkS4class{VSLCMparamContinuous}. Parameters of the continuous variables.}
  \item{paramInteger}{\linkS4class{VSLCMparamInteger}. Parameters of the integer variables.}
  \item{paramCategorical}{\linkS4class{VSLCMparamCategorical}. Parameters of the categorical variables.}
}
}
\examples{
  getSlots("VSLCMparamMixed")

}
