% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{VSLCMpartitions-class}
\alias{VSLCMpartitions-class}
\title{Constructor of \code{\linkS4class{VSLCMpartitions}} class}
\description{
\describe{
  \item{zMAP}{numeric. A vector indicating the class membership of each individual by using the MAP rule computed for the best model with its maximum likelihood estimates.}
  \item{zOPT}{numeric. Partition maximizing the integrated complete-data likelihood of the selected model.}
  \item{tik}{numeric. Fuzzy partition computed for the best model with its maximum likelihood estimates.}
}
}
\examples{
  getSlots("VSLCMpartitions")

}
