\name{VSLCMresultsCategorical-class}
\docType{class}
\alias{VSLCMresultsCategorical-class}

\title{Class \code{"VSLCMresultsCategorical"}}
\description{
This class contains the model information
}

\section{Slots}{
  \describe{
    \item{\code{data}:}{An object of class \code{\linkS4class{VSLCMdataCategorical}} }
    \item{\code{criteria}:}{An object of class \code{\linkS4class{VSLCMcriteria}}}
    \item{\code{partitions}:}{An object of class \code{\linkS4class{VSLCMpartitions}}}
    \item{\code{model}:}{An object of class \code{\linkS4class{VSLCMmodel}}}
    \item{\code{strategy}:}{An object of class \code{\linkS4class{VSLCMstrategy}}}
    \item{\code{param}:}{An object of class \code{\linkS4class{VSLCMparamCategorical}}}
  }
}

\examples{
showClass("VSLCMresultsCategorical")
}
\keyword{classes}