\name{VSLCMresultsMixed-class}
\docType{class}
\alias{VSLCMresultsMixed-class}

\title{Class \code{"VSLCMresultsMixed"}}
\description{
  This class contains the model information
}

\section{Slots}{
  \describe{
    \item{\code{data}:}{An object of class \code{\linkS4class{VSLCMdataMixed}} }
    \item{\code{criteria}:}{An object of class \code{\linkS4class{VSLCMcriteria}}}
    \item{\code{partitions}:}{An object of class \code{\linkS4class{VSLCMpartitions}}}
    \item{\code{model}:}{An object of class \code{\linkS4class{VSLCMmodel}}}
    \item{\code{strategy}:}{An object of class \code{\linkS4class{VSLCMstrategy}}}
    \item{\code{param}:}{An object of class \code{\linkS4class{VSLCMparamMixed}}}
  }
}

\examples{
  showClass("VSLCMresultsMixed")
}
\keyword{classes}