% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{VSLCMstrategy-class}
\alias{VSLCMstrategy-class}
\title{Constructor of \code{\linkS4class{VSLCMstrategy}} class}
\description{
\describe{
  \item{initModel}{numeric. Number of initialisations for the model selection algorithm.}
  \item{vbleSelec}{logical. It indicates if the selection of the variables is performed.}
  \item{paramEstim}{logical. It indicates if the parameter estimation is performed.}
  \item{parallel}{logical. It indicates  if a parallelisation is done.}
  \item{nbSmall}{numeric. It indicates the number of small EM.}
  \item{iterSmall}{numeric. It indicates the number of iteration for the small EM}
  \item{nbKeep}{numeric. It indicates the number of chains kept for the EM.}
  \item{iterKeep}{numeric. It indicates the maximum number of iteration for the EM.}
  \item{tolKeep}{numeric. It indicates the value of the difference between successive iterations of EM stopping the EM.}
}
}
\examples{
  getSlots("VSLCMstrategy")

}
