% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractors.R
\docType{methods}
\name{fitted}
\alias{fitted}
\alias{fitted,VSLCMresults-method}
\alias{fitted}
\alias{fitted,VSLCMresults-method}
\alias{fitted}
\alias{fitted.values}
\alias{fitted.values,VSLCMresults-method}
\title{Extract the estimated partition}
\usage{
\S4method{fitted}{VSLCMresults}(object)

\S4method{fitted}{VSLCMresults}(object)

\S4method{fitted.values}{VSLCMresults}(object)
}
\arguments{
\item{object}{instance of \code{\linkS4class{VSLCMresults}}.}

\item{object}{instance of \code{\linkS4class{VSLCMresults}}.}

\item{object}{instance of \code{\linkS4class{VSLCMresults}}.}
}
\description{
This function returns the partition among the observations of an instance of \code{\linkS4class{VSLCMresults}}.

This function returns the partition among the observations of an instance of \code{\linkS4class{VSLCMresults}}.

This function returns the partition among the observations of an instance of \code{\linkS4class{VSLCMresults}}.
}
