% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarSelLCM.R
\docType{data}
\name{heart}
\alias{heart}
\title{Statlog (Heart) Data Set}
\description{
This dataset is a heart disease database similar to a database already present in the repository (Heart Disease databases) but in a slightly different form.
}
\details{
12 variables are used to cluster the observations 
\itemize{
 \item{age (integer)}
 \item{sex (binary)}
 \item{chest pain type (categorical with 4 levels)}
 \item{resting blood pressure (continuous)}
 \item{serum cholestoral in mg/dl (continuous)}
 \item{fasting blood sugar > 120 mg/dl (binary)}
 \item{resting electrocardiographic results (categorical with 3 levels)}
 \item{maximum heart rate achieved (continuous)}
 \item{exercise induced angina (binary)}
 \item{the slope of the peak exercise ST segment (categorical with 3 levels)}
 \item{number of major vessels  colored by flourosopy  (categorical with 4 levels)}
 \item{thal: 3 = normal; 6 = fixed defect; 7 = reversable defect (categorical with 3 levels)}
 }
 
 1 variable define a ''true'' partition: Absence (1) or presence (2) of heart disease
}
\examples{
  data(heart)
}
\references{
UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science: http://archive.ics.uci.edu/ml/datasets/statlog+(heart)
}
\keyword{datasets}
