% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarSelLCM.R
\docType{data}
\name{heart}
\alias{heart}
\title{Statlog (Heart) Data Set}
\description{
This dataset is a heart disease database similar to a database already present in the repository (Heart Disease databases) but in a slightly different form.
}
\examples{
  data(heart)
}
\references{
Website:http://archive.ics.uci.edu/ml/datasets/statlog+(heart)
}
\keyword{datasets}
