% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\docType{methods}
\name{plot,VSLCMresultsContinuous,character-method}
\alias{plot,VSLCMresultsContinuous,character-method}
\alias{plot}
\alias{plot,}
\alias{VSLCMresultsContinuous-method}
\alias{plot,VSLCMresultsInteger,character-method}
\alias{plot}
\alias{plot,}
\alias{VSLCMresultsInteger-method}
\alias{plot}
\alias{plot,VSLCMresultsCategorical-method}
\alias{plot,VSLCMresultsMixed,character-method}
\alias{plot}
\alias{plot,VSLCMresultsMixed-method}
\title{This function draws information about an instance of \code{\linkS4class{VSLCMresultsContinuous}}, \code{\linkS4class{VSLCMresultsInteger}}, \code{\linkS4class{VSLCMresultsCategorical}} or \code{\linkS4class{VSLCMresultsMixed}}.}
\usage{
\S4method{plot}{VSLCMresultsContinuous,character}(x, y, type)

\S4method{plot}{VSLCMresultsInteger,character}(x, y, type)

\S4method{plot}{VSLCMresultsCategorical,ANY}(x)

\S4method{plot}{VSLCMresultsMixed,character}(x, y, type)
}
\arguments{
\item{object}{instance of  \code{\linkS4class{VSLCMresultsContinuous}}, \code{\linkS4class{VSLCMresultsInteger}}, \code{\linkS4class{VSLCMresultsCategorical}} or \code{\linkS4class{VSLCMresultsMixed}}..}
}
\description{
This function draws information about an instance of \code{\linkS4class{VSLCMresultsContinuous}}, \code{\linkS4class{VSLCMresultsInteger}}, \code{\linkS4class{VSLCMresultsCategorical}} or \code{\linkS4class{VSLCMresultsMixed}}.
}
