% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,VSLCMresults-method}
\title{Prediction of the cluster memberships}
\usage{
\S4method{predict}{VSLCMresults}(object, newdata, type = "probability")
}
\arguments{
\item{object}{instance of \code{\linkS4class{VSLCMresults}}.}

\item{newdata}{data.frame of the observations to classify.}

\item{type}{the type of prediction: probability of classification (probability) or the partition (partition)}
}
\value{
Returns a matrix of the probabilities of classification.
}
\description{
This function gives the probabilities of classification for new observations by using the mixture model fit with the function \code{\link{VarSelCluster}}.
}
