% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,VSLCMresults-method}
\title{Print function.}
\usage{
\S4method{print}{VSLCMresults}(x)
}
\arguments{
\item{x}{instance of \code{\linkS4class{VSLCMresults}}.}
}
\description{
This function gives the print of an instance of  \code{\linkS4class{VSLCMresults}}.
}
