\docType{data}
\name{SimulHetero}
\alias{SimulHetero}
\title{SimulHetero data}
\format{A simulated data set.

\describe{
  \item{\code{x}: the whole data set}
  \item{\code{z}: the component membership}
  \item{\code{xmiss}: data set with missing values}
}
}
\examples{
data(SimulHetero)
}
\keyword{datasets}

