% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,VSLCMresults-method}
\title{Summary function.}
\usage{
\S4method{summary}{VSLCMresults}(object)
}
\arguments{
\item{object}{instance of \code{\linkS4class{VSLCMresults}}.}
}
\description{
This function gives the summary of an instance of  \code{\linkS4class{VSLCMresults}}.
}
