% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarSelLCM.R
\name{VarSelCluster}
\alias{VarSelCluster}
\title{Variable selection and clustering.}
\usage{
VarSelCluster(x, g, vbleSelec = TRUE, crit.varsel = "BIC", initModel = 50,
  nbcores = 1, discrim = rep(1, ncol(x)), nbSmall = 250, iterSmall = 20,
  nbKeep = 50, iterKeep = 1000, tolKeep = 10^(-6))
}
\arguments{
\item{x}{data.frame. Rows correspond to observations and columns correspond to variables. Continuous variables must be "numeric", count variables must be "integer" and categorical variables must be "factor"}

\item{g}{numeric. It defines number of components}

\item{vbleSelec}{logical. It indicates if a variable selection is done}

\item{crit.varsel}{character. It defines the information criterion used for the variable selection ("AIC", "BIC" or "MICL"; only used if vbleSelec=1)}

\item{initModel}{numeric. It gives the number of initializations of the alternated algorithm maximizing the MICL criterion (only used if crit.varsel="MICL")}

\item{nbcores}{numeric.  It defines the numerber of cores used by the alogrithm}

\item{discrim}{numeric. It indicates if each variable is discrimiative (1) or irrelevant (0) (only used if vbleSelec=0)}

\item{nbSmall}{numeric. It indicates  the number of SmallEM algorithms performed for the ML inference}

\item{iterSmall}{numeric. It indicates  the number of iterations for each SmallEM algorithm}

\item{nbKeep}{numeric. It indicates the number of chains used for the final EM algorithm}

\item{iterKeep}{numeric. It indicates the maximal number of iterations for each EM algorithm}

\item{tolKeep}{numeric. It indicates the maximal gap between two successive iterations of EM algorithm which stops the algorithm}
}
\value{
Returns an instance of \linkS4class{VSLCMresultsMixed}.
}
\description{
This function performs the model selection and the maximum likelihood estimation of the Latent Class Model.
It can be used for clustering only (i.e., all the variables are assumed to be discriminative). You must specify the data to analyse (arg. x), the number of clusters (arg. g) and the option vbleSelec must be FALSE.
This function can be used for variable selection in clustering. You must specify the data to analyse (arg. x), the number of clusters (arg. g) and the option vbleSelec must be TRUE. Variable selection can be done with BIC, MICL or AIC.
}
\examples{
\dontrun{
data(iris)
res.LCM <- VarSelCluster(x, 2)
summary(res.LCM)
}
}
\references{
Marbac, M. and Sedki, M. (2017). Variable selection for model-based clustering using the integrated completed-data likelihood. Variable selection for model-based clustering using the integrated complete-data likelihood. Statistics and Computing, 27 (4), 1049–1063.

Marbac, M. and Patin, E. and Sedki, M. (2018). Variable selection for mixed data clustering: Application in human population genomics. Arxiv 1703.02293.
}
