% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{VSLCMcriteria-class}
\alias{VSLCMcriteria-class}
\title{Constructor of \code{\linkS4class{VSLCMcriteria}} class}
\description{
\describe{
  \item{loglikelihood}{numeric. Log-likelihood}
  \item{AIC}{numeric. Value of the AIC criterion.}
  \item{BIC}{numeric. Value of the BIC criterion.}
  \item{ICL}{numeric. Value of the ICL criterion.}
  \item{MICL}{numeric. Value of the MICL criterion.}
  \item{nbparam}{integer. Number of parameters.}
  \item{cvrate}{numeric.  Rate of convergence of the alternated algorithm for optimizing the MICL criterion.}
  \item{degeneracyrate}{numeric.  Rate of degeneracy for the selected model.}
  \item{discrim}{numeric.  Discriminative power of each variable.}
}
}
\examples{
  getSlots("VSLCMcriteria")

}
