% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{class}
\name{VSLCMdataCategorical-class}
\alias{VSLCMdataCategorical-class}
\title{Constructor of \code{\linkS4class{VSLCMdataCategorical}} class}
\description{
\describe{
  \item{n}{number of observations}
  \item{d}{number of variables}
  \item{data}{matrix of observations (one row = one observation)}
  \item{shortdata}{matrix of unique profils}
  \item{weightdata}{weights of profils}
  \item{modalitynames}{names of levels}
}
}
\examples{
  getSlots("VSLCMdataCategorical")

}
