% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCstr.R
\docType{class}
\name{VSLCMdataCategorical-class}
\alias{VSLCMdataCategorical-class}
\title{Constructor of [\code{\linkS4class{VSLCMdataCategorical}}] class}
\description{
\describe{
  \item{n}{nombre d'observations.}
  \item{d}{ nombre de variables.}
  \item{data}{matrix ou les facteurs orginiaux ont ete converti en numeric.}
  \item{shortdata}{matrix contenant les profils uniques.}
  \item{weightdata}{poids de chaque profil.}
  \item{modalitynames}{list contenant les noms de modalites pour chaque variable.}
}
}
\examples{
  getSlots("VSLCMdataCategorical")

}
