% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{class}
\name{VSLCMdataContinuous-class}
\alias{VSLCMdataContinuous-class}
\title{Constructor of \code{\linkS4class{VSLCMdataContinuous}} class}
\description{
\describe{ 
  \item{n}{number of observations}
  \item{d}{number of variables}
  \item{data}{matrix of observations (one row = one observation)}
  \item{notNA}{matrix of logical (1:observed, 0:unobserved)}
  \item{priors}{hyper-parameters of the prior distributions}
}
}
\examples{
  getSlots("VSLCMdataContinuous")
  

}
