% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCstr.R
\docType{class}
\name{VSLCMdataInteger-class}
\alias{VSLCMdataInteger-class}
\title{Constructor of [\code{\linkS4class{VSLCMdataInteger}}] class}
\description{
\describe{
  \item{n}{nombre d'observations.}
  \item{d}{ nombre de variables.}
  \item{data}{matrix ou les colonnes sont numeric et correspondent aux donees.}
  \item{notNA}{matrix of logical valant 1 si la realisation est observee et 0 sinon.}
  \item{priors}{valeur des priors pour chaque variable (en ligne).}
}
}
\examples{
  getSlots("VSLCMdataInteger")

}
