% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCstr.R
\docType{class}
\name{VSLCMdataMixed-class}
\alias{VSLCMdataMixed-class}
\title{Constructor of [\code{\linkS4class{VSLCMdataMixed}}] class
\describe{
\item{n}{le nombre d'observations.}
\item{d}{le nombre de variables.} 
\item{withContinuous}{boolien qui indique si variables continues ou pas.}
\item{withInteger}{boolien qui indique si variables entieres ou pas.}
\item{withCategorica}{boolien qui indique si variables categorielles ou pas.} 
\item{dataContinuous}{objet de la calsse VSLCMdataContinuous pour la partie continue des donnees.}
\item{dataInteger}{objet de la classe VSLCMdataInteger pour la partie entiere des donnees.}
\item{dataCategorical}{objet de la classe VSLCMdataCategorical pour la partie categorielle des donnees.}
\item{var.names}{caracteres contenant les noms des variables.} 
}}
\description{
Constructor of [\code{\linkS4class{VSLCMdataMixed}}] class
\describe{
\item{n}{le nombre d'observations.}
\item{d}{le nombre de variables.} 
\item{withContinuous}{boolien qui indique si variables continues ou pas.}
\item{withInteger}{boolien qui indique si variables entieres ou pas.}
\item{withCategorica}{boolien qui indique si variables categorielles ou pas.} 
\item{dataContinuous}{objet de la calsse VSLCMdataContinuous pour la partie continue des donnees.}
\item{dataInteger}{objet de la classe VSLCMdataInteger pour la partie entiere des donnees.}
\item{dataCategorical}{objet de la classe VSLCMdataCategorical pour la partie categorielle des donnees.}
\item{var.names}{caracteres contenant les noms des variables.} 
}
}
\examples{
  getSlots("VSLCMdataMixed")

}
