\name{VSLCMmodel-class}
\docType{class}
\alias{VSLCMmodel-class}

\title{Class \code{"VSLCMmodel"}}
\description{
This class contains the model information
}



\section{Slots}{
  \describe{
    \item{\code{g}:}{Number of components}
    \item{\code{omega}:}{Boolean vector indicating if each variable is relevant (1) or not (0) to the clustering}
  }
}


\examples{
showClass("VSLCMmodel")
}

\keyword{classes}
