% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSLCMGrlClasses.R
\docType{class}
\name{VSLCMparamContinuous-class}
\alias{VSLCMparamContinuous-class}
\title{Constructor of [\code{\linkS4class{VSLCMparamContinuous}}] class}
\description{
\describe{
  \item{pi}{numeric. Proportions of the mixture components.}
  \item{mu}{matrix. Mean for each component (column) and each variable (row).}
  \item{sd}{matrix. Standard deviation for each component (column) and each variable (row).}
}
}
\examples{
  getSlots("VSLCMparamContinuous")

}
