% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{class}
\name{VSLCMresultsCategorical-class}
\alias{VSLCMresultsCategorical-class}
\title{Constructor of \code{\linkS4class{VSLCMresultsCategorical}} class}
\description{
\describe{
  \item{data}{\linkS4class{VSLCMdataCategorical}. Results relied to the data.}
  \item{criteria}{\linkS4class{VSLCMcriteria}. Results relied to the information criteria.}
  \item{partitions}{\linkS4class{VSLCMpartitions}. Results relied to the partitions.}
  \item{model}{\linkS4class{VSLCMmodel}. Results relied to the selected model.}
  \item{strategy}{\linkS4class{VSLCMstrategy}. Results relied to the tune parameters.}
  \item{param}{\linkS4class{VSLCMparamCategorical}. Results relied to the parameters.}
}
}
\examples{
  getSlots("VSLCMresultsCategorical")

}
