% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSLCMGrlClasses.R
\docType{class}
\name{VSLCMresultsMixed-class}
\alias{VSLCMresultsMixed-class}
\title{Constructor of [\code{\linkS4class{VSLCMresultsMixed}}] class}
\description{
\describe{
  \item{data}{\linkS4class{VSLCMdataMixed}. Results relied to the data.}
  \item{criteria}{\linkS4class{VSLCMcriteria}. Results relied to the information criteria.}
  \item{partitions}{\linkS4class{VSLCMpartitions}. Results relied to the partitions.}
  \item{model}{\linkS4class{VSLCMmodel}. Results relied to the selected model.}
  \item{strategy}{\linkS4class{VSLCMstrategy}. Results relied to the tune parameters.}
  \item{param}{\linkS4class{VSLCMparamMixed}. Results relied to the parameters.}
}
}
\examples{
  getSlots("VSLCMresultsMixed")

}
